Add files via upload
# Node Distance Tool for Godot

The Node Distance Tool is an addon for Godot Engine that allows users to select two 3D nodes and measure the distance between them. It creates a line connecting the nodes and displays the measured distance in the scene. This addon is useful for developers who need to measure distances between objects during game or simulation development in Godot.

## Features
- Select two 3D nodes and measure the distance between them.
- Visualize a line connecting the two nodes.
- Display the rounded distance in meters between the nodes.
- Shortcut "Q" for easy selection and measurement.
- Option to reset measurements (remove all lines and labels) with a dedicated button.

## How to Use
1. Install the addon in your project's addons folder in Godot.
2. Enable the addon in "Project > Project Settings > Plugins".
3. In the 3D scene editor, select the first node and press "Q".
4. Select the second node and press "Q" again to measure the distance.
5. Use the reset button to remove the measurement lines and labels, if necessary.

## Requirements
- Godot Engine version 4.0 or higher.

## Notice
This addon cannot be sold separately as a standalone product. It is free and must remain free. However, it can be included and used in other products and projects.

## License
This project is licensed under the terms described in the LICENSE file.